Adapter Completed
Trying this one more time, this is me copying the code I had completed
from the previous (broken) repo into this one. Hopefully nothing breaks
this time...

In this step, I completed the implementation of an adapter for the
RecyclerView that will display the movie posters and titles. I was
able to do this by creating a custom ViewHolder implementation, which
stores the values as fields to be retrieved and populated by the
adapter. I also created a method in MainActivity that populates some
dummy data so this may be tested before I begin trying to use Picasso
to do the ImageView part for the poster as well. Seems good so far! :)