Added an OnClickListener
Implemented RecyclerView.OnClickListener with the dummy data in my
adapter, since this is where I'll need to construct the explicit intent
(and extras) that will be required to call up the detail view, which
is yet to be implemented. I also added a Silicon Valley reference, can
you find it? :)