Implemented DetailActivity
I now have an (admittedly basic) implementation of invoking
DetailActivity and passing the data from the adapter to the new
Activity! :) The only part left now is to create a functionality to
call the Movies API and convert the JSON into strings that I can work
with. Should be fun!