Implemented Hidden ApiKey
With this update, I figured out how to add a resource file (res/values/
apikey.xml) with a string attribute named api_key that can be accessed
from within the program, and added that file to the root .gitignore
file so that it won't be pushed into the repo. This commit should test
that, by excluding that res file by default.