Debug Hell
Left all my Log messages in this commit for hilarity's sake. It turns
out, the bug I was getting was a bug with the emulator and I never
needed to add all this debug. How hilarious. Also implemented some
AsyncTaskLoader bollocks, but idk if this belongs in MainActivity or
in MovieDbHelper. Hopefully my mentor can provide some insight.