JSON Parsing Finished
This commit sees the completion of the JSON parsing utility, and
subsequently the removal of all debug/test bits of code. We're very
close to spec with this update, the last part is just to implement
the ability to change the sort with a setting. I also began implementing
the preferences/settings menu that will control that sort option, but
haven't gotten all the way there yet.