Feature Complete!
At this point, I've overridden the menu in MainActivity with an option
that opens a SettingFragment, and I'm using SharedPreferences to store
the selected setting so it can be referenced when loading from the API.
This should be feature-complete according the rubric and the project
specification, so I'm submitting this commit! Yay!