Revision After Submission
So, my first submission was rejected with some very constructive
comments. This commit contains a fix for the bug that was identified
which caused me to fall short of spec, as well as fixes/enhancements
for all the other feedback that was included in my project review!
Thanks again to the reviewer who submitted such great feedback to me!